Moved VMS files into VMS directory
