+++ python/lsst/ts/cRIOpy/VMS/Bars.py
